CBD-6348: [BP] Update for C++20 compatibility
This is a cherry-pick from trinity to neo

Reviewed-on: https://review.couchbase.org/c/platform/+/234955
Well-Formed: Restriction Checker
Tested-by: Trond Norbye <trond.norbye@couchbase.com>

+++ src/backtrace.cc
@@ -12,6 +12,8 @@
 #include <cinttypes>
 
 #if defined(WIN32)
+#include <cstdlib>
+#include <system_error>
 #include <windows.h>
 #include <Dbghelp.h>
 #else
@@ -136,7 +138,8 @@ void print_backtrace_frames(const boost::stacktrace::stacktrace& frames,
     for (size_t ii = 0; ii < frames.size(); ii++) {
         // Fixed-sized buffer; possible that description will be cropped.
         char msg[300];
-        describe_address(msg, sizeof(msg), ii, frames[ii].address());
+        describe_address(
+                msg, sizeof(msg), static_cast<int>(ii), frames[ii].address());
         callback(msg);
     }
 }
